requirements.txt
# ═══════════════════════════════════════════════════════════
# AetherWatch — Requirements
# Python 3.11+ | pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════

# ── Core Web Framework ──────────────────────────────────────
streamlit==1.32.0
streamlit-folium==0.20.0
streamlit-autorefresh==1.0.1

# ── Mapping ─────────────────────────────────────────────────
folium==0.17.0

# ── Computer Vision & YOLO ──────────────────────────────────
ultralytics==8.1.27
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# ── Data & HTTP ─────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
requests==2.31.0

# ── Config & Env ────────────────────────────────────────────
python-dotenv==1.0.1

# ── Logging ─────────────────────────────────────────────────
loguru==0.7.2

# ── Caching ─────────────────────────────────────────────────
cachetools==5.3.3
